requirements.txt
streamlit>=1.28.0
streamlit-quill>=0.0.3
Pillow>=9.0.0
python-dateutil>=2.8.0
pydantic>=2.0.0
# ai_tools will be installed from local wheel file
